+++ sshd-sftp/src/main/java/org/apache/sshd/sftp/server/SftpUnsupportedAttributePolicyProvider.java
@@ -22,6 +22,10 @@ package org.apache.sshd.sftp.server;
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public interface SftpFileSystemAccessorManager extends SftpFileSystemAccessorProvider {
-    void setFileSystemAccessor(SftpFileSystemAccessor accessor);
+@FunctionalInterface
+public interface SftpUnsupportedAttributePolicyProvider {
+    /**
+     * @return The {@link UnsupportedAttributePolicy} to use if failed to access some local file attributes
+     */
+    UnsupportedAttributePolicy getUnsupportedAttributePolicy();
 }