+++ sshd-sftp/src/test/java/org/apache/sshd/sftp/client/extensions/helpers/SpaceAvailableExtensionImplTest.java
@@ -71,9 +71,7 @@ public class SpaceAvailableExtensionImplTest extends AbstractSftpClientTestSuppo
         sshd.setSubsystemFactories(Collections.singletonList(new SftpSubsystemFactory() {
             @Override
             public Command createSubsystem(ChannelSession channel) throws IOException {
-                return new SftpSubsystem(
-                        resolveExecutorService(),
-                        getUnsupportedAttributePolicy(), getFileSystemAccessor(), getErrorStatusDataHandler()) {
+                return new SftpSubsystem(this) {
                     @Override
                     protected SpaceAvailableExtensionInfo doSpaceAvailable(int id, String path) throws IOException {
                         if (!queryPath.equals(path)) {